pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dq-copilot"
version = "0.1.0"
description = "CLI AI agent that profiles tables, detects data quality issues, and generates dbt-style tests"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Data Quality Team"}
]
dependencies = [
    "click>=8.0.0",
    "pandas>=1.3.0",
    "pyarrow>=6.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "rich>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
dq-copilot = "dq_copilot.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dq_copilot*"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
target-version = "py38"
